refactor(converters hci, synth): move specific structs out of types.rs
Have a rust code file for each parser that includes the types that only occur in that parser: that way we can also distinguish between how structures should be expected in each parser. This approach is supposed to improve maintainability in the longterm.

refactor(catplus-common): rename `types.rs` and makes submodule `types`
