+++ src/catplus-common/src/models/core.rs
@@ -5,198 +5,14 @@
 use crate::{
     graph::{
         insert_into::{InsertIntoGraph, Link},
-        namespaces::{alloproc, allocom, allohdf, alloqual, allores, cat, obo, purl, qudt, schema},
+        namespaces::{alloqual, allores, cat, obo, purl, qudt, schema},
     },
-    models::enums::{ActionName, Unit},
+    models::enums::Unit,
 };
 use anyhow;
 use serde::{Deserialize, Serialize};
-use sophia::{
-    api::ns::{rdf, xsd},
-    inmem::graph::LightGraph,
-};
-use sophia_api::{
-    graph::MutableGraph,
-    term::{SimpleTerm, Term},
-};
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Campaign {
-    pub campaign_name: String,
-    pub description: String,
-    #[serde(rename = "objective")]
-    pub generic_objective: String,
-    pub campaign_class: String,
-    #[serde(rename = "type")]
-    pub campaign_type: String,
-    pub reference: String,
-    pub has_objective: Option<Objective>,
-    pub has_batch: Batch,
-    pub has_chemical: Option<Vec<Chemical>>,
-}
-
-impl InsertIntoGraph for Campaign {
-    fn insert_into(&self, graph: &mut LightGraph, iri: SimpleTerm) -> anyhow::Result<()> {
-        for (pred, value) in [
-            (rdf::type_, &cat::Campaign.as_simple() as &dyn InsertIntoGraph),
-            (schema::name, &self.campaign_name.as_simple()),
-            (schema::description, &self.description.as_simple()),
-            (cat::genericObjective, &self.generic_objective.as_simple()),
-            (cat::campaignClass, &self.campaign_class.as_simple()),
-            (cat::campaignType, &self.campaign_type.as_simple()),
-            (allores::AFR_0002764, &self.reference.as_simple()),
-            (cat::hasObjective, &self.has_objective),
-            (cat::hasBatch, &self.has_batch),
-            (cat::hasChemical, &self.has_chemical),
-        ] {
-            value.attach_into(
-                graph,
-                Link { source_iri: iri.clone(), pred: pred.as_simple(), target_iri: None },
-            )?;
-        }
-        Ok(())
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Objective {
-    pub criteria: String,
-    pub condition: String,
-    pub description: String,
-    pub objective_name: String,
-}
-
-impl InsertIntoGraph for Objective {
-    fn insert_into(&self, graph: &mut LightGraph, iri: SimpleTerm) -> anyhow::Result<()> {
-        for (pred, value) in [
-            (rdf::type_, &obo::IAO_0000005.as_simple()),
-            (schema::name, &self.objective_name.as_simple()),
-            (schema::description, &self.description.as_simple()),
-            (cat::criteria, &self.criteria.as_simple()),
-            (allocom::AFC_0000090, &self.condition.as_simple()),
-        ] {
-            value.attach_into(
-                graph,
-                Link { source_iri: iri.clone(), pred: pred.as_simple(), target_iri: None },
-            )?;
-        }
-        Ok(())
-    }
-}
-
-#[derive(Deserialize)]
-pub struct CampaignWrapper {
-    #[serde(rename = "hasCampaign")]
-    pub has_campaign: Campaign,
-}
-impl InsertIntoGraph for CampaignWrapper {
-    fn insert_into(&self, graph: &mut LightGraph, iri: SimpleTerm) -> anyhow::Result<()> {
-        self.has_campaign.insert_into(graph, iri)
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Batch {
-    #[serde(rename = "batchID")]
-    pub batch_id: String,
-    #[serde(rename = "Actions")]
-    pub actions: Option<Vec<Action>>,
-    pub batch_name: Option<String>,
-    pub reaction_type: Option<String>,
-    pub reaction_name: Option<String>,
-    pub optimization_type: Option<String>,
-    pub link: Option<String>,
-}
-
-impl InsertIntoGraph for Batch {
-    fn insert_into(&self, graph: &mut LightGraph, iri: SimpleTerm) -> anyhow::Result<()> {
-        for (pred, value) in [
-            (rdf::type_, &cat::Batch.as_simple() as &dyn InsertIntoGraph),
-            (purl::identifier, &self.batch_id.as_simple()),
-            (schema::name, &self.batch_name.as_ref().clone().map(|s| s.as_simple())),
-            (allohdf::HardLink, &self.link.as_ref().clone().map(|s| s.as_simple())),
-            (cat::reactionType, &self.reaction_type.as_ref().clone().map(|s| s.as_simple())),
-            (cat::reactionName, &self.reaction_name.as_ref().clone().map(|s| s.as_simple())),
-            (
-                cat::optimizationType,
-                &self.optimization_type.as_ref().clone().map(|s| s.as_simple()),
-            ),
-        ] {
-            value.attach_into(
-                graph,
-                Link { source_iri: iri.clone(), pred: pred.as_simple(), target_iri: None },
-            )?;
-        }
-
-        // NOTE: for actions, the direction is reversed (action hasbatch batch)
-        if let Some(actions) = &self.actions {
-            for action in actions {
-                let action_uri = action.get_uri();
-                graph.insert(&action_uri, cat::hasBatch.as_simple(), iri.clone())?;
-                action.insert_into(graph, action_uri)?;
-            }
-        }
-
-        Ok(())
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Action {
-    pub action_name: ActionName,
-    pub start_time: String,
-    pub ending_time: String,
-    pub method_name: String,
-    pub equipment_name: String,
-    pub sub_equipment_name: String,
-    #[serde(flatten)]
-    pub has_plate: Option<Plate>,
-    pub speed_shaker: Option<Observation>,
-    pub has_well: Option<Vec<Well>>,
-    pub dispense_state: Option<String>,
-    pub dispense_type: Option<String>,
-    pub has_sample: Option<Sample>,
-    pub speed_tumble_stirrer: Option<Observation>,
-    pub temperature_tumble_stirrer: Option<Observation>,
-    pub temperature_shaker: Option<Observation>,
-    pub pressure_measurement: Option<Observation>,
-    pub vacuum: Option<Observation>,
-}
-
-impl InsertIntoGraph for Action {
-    fn insert_into(&self, graph: &mut LightGraph, iri: SimpleTerm) -> anyhow::Result<()> {
-        for (pred, value) in [
-            (rdf::type_, &self.action_name.iri().as_simple() as &dyn InsertIntoGraph),
-            (allores::AFX_0000622, &(self.start_time.as_str() * xsd::dateTime).as_simple()),
-            (allores::AFR_0002423, &(self.ending_time.as_str() * xsd::dateTime).as_simple()),
-            (allores::AFR_0001606, &self.method_name.as_simple()),
-            (allores::AFR_0001723, &self.equipment_name.as_simple()),
-            (cat::subEquipmentName, &self.sub_equipment_name.as_simple()),
-            (cat::speedInRPM, &self.speed_shaker),
-            (cat::temperatureTumbleStirrerShape, &self.temperature_tumble_stirrer),
-            (cat::speedTumbleStirrerShape, &self.speed_tumble_stirrer),
-            (cat::vacuum, &self.vacuum),
-            (cat::temperatureShakerShape, &self.temperature_shaker),
-            (alloproc::AFP_0002677, &self.pressure_measurement),
-            (cat::hasSample, &self.has_sample),
-            (cat::hasWell, &self.has_well),
-            (cat::hasPlate, &self.has_plate),
-            (alloqual::AFQ_0000111, &self.dispense_state.as_ref().clone().map(|s| s.as_simple())),
-            (cat::dispenseType, &self.dispense_type.as_ref().clone().map(|s| s.as_simple())),
-        ] {
-            value.attach_into(
-                graph,
-                Link { source_iri: iri.clone(), pred: pred.as_simple(), target_iri: None },
-            )?;
-        }
-
-        Ok(())
-    }
-}
+use sophia::{api::ns::rdf, inmem::graph::LightGraph};
+use sophia_api::term::{SimpleTerm, Term};
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
