+++ src/catplus-common/src/models/mod.rs
@@ -1,7 +1,10 @@
 pub mod agilent;
+pub mod core;
 pub mod enums;
 pub mod types;
+
 // Re-export all models;
 pub use agilent::*;
+pub use core::*;
 pub use enums::*;
 pub use types::*;
