+++ src/catplus-common/src/models/types/mod.rs
@@ -1,0 +1,2 @@
+pub mod hci;
+pub mod synth;
